requirements.txt
attr==0.3.2
ConfigParser==7.2.0
cryptography==46.0.3
docutils==0.22.3
filelock==3.20.0
HTMLParser==0.0.2
ipython==9.7.0
ipywidgets==8.1.8
keyring==25.6.0
protobuf==6.33.1
pyOpenSSL==25.3.0
redis==7.0.1
thread==2.0.5
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
requests==2.32.5
python-dotenv==1.2.1
black==25.11.0
flake8==7.3.0
pylint==4.0.3
sqlalchemy==2.0.44
psycopg2-binary==2.9.11